\name{climatedata}
\docType{data}
\alias{climatedata}
\title{Raw Climate Data Required for Calculating Evapotranspiration}
\description{
  This data set contains the raw climate data including the variables required for calculating evapotranspiration in function \code{Evapotranspiration} over the observation period between 1/3/2001 and 08/31/2004 at the Kent Town station in Adelaide, Australia.
}
\usage{climatedata}
\format{A list containing 11688 obserations of 30 variables}
\source{Bureau of Meteorology, Kent Town, Adelaide, Australia}
\keyword{datasets}
