\name{E_OBS}
\docType{data}
\alias{E_OBS}
\title{Observed Class-A Pan Evaporation}
\description{
  This data set contains the Class-A pan evaporation observed over the period between 1/3/2001 and 31/8/2004 at the Kent Town station in Adelaide, Australia.
}
\usage{E_OBS}
\format{A list containing 48 obserations of 5 variables}
\source{Bureau of Meteorology, Kent Town, Adelaide, Australia}
\keyword{datasets}
