\name{ET.default}
\alias{ET.default}

\title{ET Formulations(Internal)
}
\description{
And internal function to select the ET formulation to use when \code{\link{ET}} is called, without user-specified ET model. 
See \code{\link{ET}} for details. 
}
\author{
Danlu Guo
}

\seealso{
\code{\link{ET}}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{evapotranspiration}
\keyword{open-water evaporation}
\keyword{potential evapotranspiration}
\keyword{internal}
% __ONLY ONE__ keyword per line
