\name{Radiation}
\alias{Radiation}

\title{Radiation Component for Morton Models (Internal)
}
\description{
Estimating the radiation component for using Morton models to estimate evapotranspiration.
}

\value{
This function returns a object named \code{variables} which contains variable values for implementing Morton CRAE and CRWE formulations 

}
\references{
MCMAHON, T., PEEL, M., LOWE, L., SRIKANTHAN, R. & MCVICAR, T. 2012. Estimating actual, potential, reference crop and pan evaporation using standard meteorological data: a pragmatic synthesis. Hydrology and Earth System Sciences Discussions, 9, 11829-11910.

MORTON, F. I. 1983. Operational estimates of areal evapotranspiration and their significance to the science and practice of hydrology. Journal of Hydrology, 66, 1-76.
}
\author{
Danlu Guo
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{radiation}
\keyword{Morton}
\keyword{evapotranspiration}
\keyword{internal}
% __ONLY ONE__ keyword per line

