\name{ReadInput_InterpMissing}
\alias{ReadInput_InterpMissing}

\title{Interpolation for missing raw data (Internal)
}
\description{
Internal function which performs interpolation for missing entries in the raw input climate data. See \code{\link{ReadInputs}} for details.
}

\author{
Danlu Guo
}
\seealso{
\code{\link{ReadInputs}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ReadInputs}
\keyword{data pre-processing}
\keyword{evapotranspiration}
\keyword{internal}
% __ONLY ONE__ keyword per line

