\name{constants}
\docType{data}
\alias{constants}
\title{Constants Required for Calculating Evapotranspriation}
\description{
  This data set contains the constants required for calculating evapotranspiration in function \code{Evapotranspiration} for Kent Town station in Adelaide, Australia.
}
\usage{climatedata}
\format{A list containing 41 constant values}
\source{various references}
\keyword{datasets}
