\name{data}
\docType{data}
\alias{data}
\title{Processed Climate Data Required for Calculating Evapotranspiration}
\description{
  This data set contains the processed climate data including the variables required for calculating evapotranspiration in function \code{Evapotranspiration} over the observation period between 1/3/2001 and 28/2/2005 at the Kent Town station in Adelaide, Australia.
}
\usage{data}
\format{A list containing 20 climate variables}
\source{Bureau of Meteorology, Kent Town, Adelaide, Australia}
\keyword{datasets}
