\name{Radiation}
\alias{Radiation}

\title{Radiation Component for Morton Models (Internal)
}
\description{
Estimating the radiation component for using Morton models to estimate evapotranspiration.
}

\value{
This function returns a object named \code{variables} which contains variable values for implementing Morton CRAE and CRWE formulations 

}
\references{
McMahon, T., Peel, M., Lowe, L., Srikanthan, R. & McVicar, T. 2012. \emph{Estimating actual, potential, reference crop and pan evaporation using standard meteorological data: a pragmatic synthesis}. Hydrology and Earth System Sciences Discussions, 9, 11829-11910.

Morton, F.I. 1983, \emph{Operational estimates of areal evapotranspiration and their significance to the science and practice of hydrology}. Journal of Hydrology, vol. 66, no. 1-4, pp. 1-76.

Morton, F.I. 1983, \emph{Operational estimates of lake evaporation}. Journal of Hydrology, vol. 66, no. 1-4, pp. 77-100.
}
\author{
Danlu Guo
}
\seealso{
\code{\link{defaultconstants}},\code{\link{constants}},\code{\link{ET.MortonCRAE}},\code{\link{ET.MortonCRWE}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{radiation}
\keyword{Morton}
\keyword{evapotranspiration}
\keyword{internal}
% __ONLY ONE__ keyword per line

