\name{ReadInput_InterpAbnormal}
\alias{ReadInput_InterpAbnormal}

\title{Fix abnormal values in raw climate data (Internal)
}
\description{
This is an internal model used to interpolate and replace the abnormal values in raw climate data, please see \code{ReadInputs} for details.
}

\author{
Danlu Guo
}
\seealso{
\code{\link{ReadInputs}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{evapotranspiration}
\keyword{internal}
\keyword{ReadInputs}
\keyword{data pre-processing}
% __ONLY ONE__ keyword per line

