\name{ReadInput_InterpMissing}
\alias{ReadInput_InterpMissing}

\title{Interpolation for missing entries in raw climate data (Internal)
}
\description{
This is an internal model used to fill the missing entries in raw climate data, please see \code{ReadInputs} for details.
}

\author{
Danlu Guo
}
\seealso{
\code{\link{ReadInputs}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{evapotranspiration}
\keyword{internal}
\keyword{ReadInputs}
\keyword{data pre-processing}
% __ONLY ONE__ keyword per line

